backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pypdf==4.3.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Celery
celery==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI and ML
openai==1.10.0
anthropic==0.18.0
google-generativeai==0.8.3
faiss-cpu==1.8.0
numpy==1.26.3
tenacity==8.2.3

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
httpx==0.25.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
